Adding sh-mode as mode for void-packages templates
I'm using Void Linux, a Linux distribution with a minimal init system and a very
interesting package manager.
The base of the package manager, the packages, are defined in a git
repo. so called `srcpkgs` are built using a `template` file. It even allows for `patches`
to be applied on the `sourcecode`.

Anyways, the `template` files are basically `sh`-files with special functions to
be implemented/overwritten. So I'd like to recognize them and activate
`sh`-mode.

This adds a function to projectile-after-switch-project-hook which
activates sh-mode for "template" files and reloads the first opened buffer.